+++ ninja-core/src/main/java/ninja/bodyparser/BodyParserEngine.java
@@ -24,6 +24,8 @@ public interface BodyParserEngine {
      * Invoke the parser and get back a Java object populated
      * with the content of this request.
      * 
+     * MUST BE THREAD SAFE TO CALL!
+     * 
      * @param context The context
      * @param classOfT The class we expect
      * @return The object instance populated with all values from raw request
@@ -32,6 +34,8 @@ public interface BodyParserEngine {
     
     /**
      * The content type this BodyParserEngine can handle
+     * 
+     * MUST BE THREAD SAFE TO CALL!
      * 
      * @return the content type. this parser can handle - eg. "application/json"
      */