+++ ninja-core/src/main/java/ninja/bodyparser/BodyParserEngineManagerImpl.java
@@ -16,7 +16,6 @@
 
 package ninja.bodyparser;
 
-import java.util.HashMap;
 import java.util.Map;
 
 import com.google.common.collect.ImmutableMap;