+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -4,6 +4,7 @@ Version X
  * bump to devbliss.doctest 0.6.5 (ra)
  * free marker templates changes no longer cause ninja restart (makotan)
  * freemarker templates performance optimizations (ra)
+ * some improvements to harden thread-safe behavior of ninja (ra)
 
 
 Version 1.4.4